refactor: extract connection drawing
